Update types for PrimeNumbersService
Updated references of Long to Int in the PrimeNumbersService and related
classes.

+++ src/main/scala/prime/numbers/PrimeNumbersServiceImpl.scala
@@ -4,8 +4,8 @@ import com.prime.numbers.thrift._
 import com.twitter.inject.Logging
 import com.twitter.util.Future
 
-class PrimeNumbersServiceImpl extends PrimeNumbersService.MethodPerEndpoint  with Logging{
-  override def primeNumbersUntil(number: Long): Future[collection.Seq[Long]] = {
+class PrimeNumbersServiceImpl extends PrimeNumbersService.MethodPerEndpoint  with Logging {
+  override def primeNumbersUntil(number: Int): Future[collection.Seq[Int]] = {
     info(s"Request for: $number")
     Future.value(Seq(number))
   }
