+++ src/main/thrift/prime_numbers_service.thrift
@@ -1,6 +1,6 @@
 namespace scala com.prime.numbers.thrift
 
 service PrimeNumbersService {
-    list<i64> primeNumbersUntil(1: i64 number)
+    list<i32> primeNumbersUntil(1: i32 number)
 
 }
